requirements.txt
python-pptx==1.0.2
lxml>=4.9.0
